kindex: GamePath - движение по контрольным точкам
git-svn-id: svn://localhost/steel.new/trunk@99 99e027b2-380c-0410-a737-a5784e1a7bde

+++ steel/trunk/code/physic/physic_engine_3d.cpp
@@ -21,7 +21,7 @@ bool PhysicEngine3D::process(steel::time globalTime, steel::time time)
 		if(kind == PhysicInterface::custom)
 		{
 			v3 p = o.getPosition();
-			p += o.getVelocity()*time;
+			p += o.getVelocity()*(float)time;
 			o.setPosition(p);
 		}
 		if(kind == PhysicInterface::uni)
